Refactor code.
Move each model to separate file
Switched ActiveSupport to 3.0.0
Removed ActiveSupport monkeypatch
Moved URI patch to it's own file
