+++ lib/lighthouse/changeset.rb
@@ -1,0 +1,5 @@
+module Lighthouse
+  class Changeset < Base
+    site_format << '/projects/:project_id'
+  end
+end
