+++ lib/lighthouse/membership.rb
@@ -1,0 +1,8 @@
+module Lighthouse
+  class Membership < Base
+    site_format << '/users/:user_id'
+    def save
+      raise Error, "Cannot modify memberships from the API"
+    end
+  end
+end
