+++ lib/lighthouse/message.rb
@@ -1,0 +1,5 @@
+module Lighthouse
+  class Message < Base
+    site_format << '/projects/:project_id'
+  end
+end
