+++ lib/lighthouse/user.rb
@@ -1,0 +1,7 @@
+module Lighthouse
+  class User < Base
+    def memberships(options = {})
+      Membership.find(:all, :params => {:user_id => id})
+    end
+  end
+end
